simple_scanner/requirements.txt
# Dépendances minimales pour le scanner GestSport

# Base de données PostgreSQL
psycopg2-binary>=2.9.0

# Client HTTP pour l'API
requests>=2.28.0

# Traitement d'images et QR codes
opencv-python>=4.8.0
Pillow>=9.0.0
pyzbar>=0.1.8

# Interface graphique (optionnel)
# tkinter  # Inclus dans Python standard

# Logging et utilitaires
python-dateutil>=2.8.0

# Sécurité et configuration
cryptography>=3.4.0

# Performance et monitoring
psutil>=5.8.0
